refactor(mcpbridge): isolate upstream runtime planning
