+++ Sources/XcodeMCPKit/UpstreamTopologySnapshot.swift
