+++ Sources/XcodeMCPProxyKit/ProxyServer+Startup.swift
